refactor: move database admin logic into service (#4893)
